+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 // 作成したウィジェットのインポート
-import 'hiraku/test.dart';
+import 'hiraku/EventRegister.dart';
 
 void main() {
   runApp(EventRegisterWidget());
